Refactor: move loss function def
